fix: vusion install

+++ src/lib/utils/index.ts
@@ -31,23 +31,31 @@ export const getAppConfig = function (pagePath: string): AppConfig {
   return appConfig;
 };
 
-import * as os from 'os';
-import * as shell from 'shelljs';
+import * as os from "os";
+import * as shell from "shelljs";
 
-const { spawn } = require('child_process');
+const { spawn } = require("child_process");
 
 function exec(command: string, options = {}) {
-    // const command = args.join(' ');
-    console.log(command);
+  // const command = args.join(' ');
+  console.log(command);
 
-    return new Promise((resolve, reject) => {
-        const result = spawn(command, Object.assign({
-            shell: true,
-            stdio: 'inherit',
-        }, options));
-        result.on('error', reject);
-        result.on('close', (code: number) => (code === 0 ? resolve(code) : reject()));
-    });
+  return new Promise((resolve, reject) => {
+    const result = spawn(
+      command,
+      Object.assign(
+        {
+          shell: true,
+          stdio: "inherit",
+        },
+        options
+      )
+    );
+    result.on("error", reject);
+    result.on("close", (code: number) =>
+      code === 0 ? resolve(code) : reject()
+    );
+  });
 }
 
 /**
@@ -59,27 +67,40 @@ function exec(command: string, options = {}) {
  * @param name If you want to rename. Defaults to package@version
  * @param clearExisting
  */
-export async function npmDownload(info: {
-    registry?: string, name: string, version?: string,
-}, dir: string, name?: string, clearExisting?: boolean) {
-    const registry = info.registry || 'https://registry.npmjs.org';
-    const version = info.version || 'latest';
+export async function npmDownload(
+  info: {
+    registry?: string;
+    name: string;
+    version?: string;
+  },
+  dir: string,
+  name?: string,
+  clearExisting?: boolean
+) {
+  const registry = info.registry || "https://registry.npmjs.org";
+  const version = info.version || "latest";
 
-    name = name || info.name.replace(/\//, '__');
-    const dest = path.join(dir, name);
-    if (fs.existsSync(dest)) {
-        if (clearExisting)
-            fs.removeSync(dest);
-        else
-            return dest;
-    }
+  name = name || info.name.replace(/\//, "__");
+  const dest = path.join(dir, name);
+  if (fs.existsSync(dest)) {
+    if (clearExisting) fs.removeSync(dest);
+    else return dest;
+  }
 
-    const temp = path.resolve(os.tmpdir(), name + '-' + new Date().toJSON().replace(/[-:TZ]/g, '').slice(0, -4));
-    await fs.ensureDir(temp);
+  const temp = path.resolve(
+    os.tmpdir(),
+    name +
+      "-" +
+      new Date()
+        .toJSON()
+        .replace(/[-:TZ]/g, "")
+        .slice(0, -4)
+  );
+  await fs.ensureDir(temp);
 
-    await exec(`npm i ${info.name}@${version}`, { cwd: temp })
-    await fs.move(path.join(temp, 'node_modules', info.name), dest);
-    fs.removeSync(temp);
+  await exec(`cd ${temp} && npm i ${info.name}@${version}`);
+  await fs.move(path.join(temp, "node_modules", info.name), dest);
+  fs.removeSync(temp);
 
-    return dest;
+  return dest;
 }
